using class of exceptions

+++ opinions_app/api_views.py
@@ -1,6 +1,7 @@
 from flask import Response, json, request
 
 from . import app, db
+from .error_handlers import InvalidAPIUsage
 from .models import Opinion
 from .views import random_opinion
 
@@ -10,14 +11,17 @@ def get_random_opinion():
     """Выводит случаный отзыв о фильме"""
     opinion = random_opinion()
 
-    json_string = json.dumps({"opinion": opinion}, ensure_ascii=False)
+    if opinion:
+        json_string = json.dumps({"opinion": opinion}, ensure_ascii=False)
 
-    response = Response(
-        json_string,
-        status=201,
-        content_type="application/json; charset=utf-8",
-    )
-    return response
+        response = Response(
+            json_string,
+            status=201,
+            content_type="application/json; charset=utf-8",
+        )
+        return response
+
+    raise InvalidAPIUsage("В базе данных нет мнений", 404)
 
 
 @app.route("/api/opinions/", methods=["GET"])
@@ -40,7 +44,13 @@ def get_opinions():
 @app.route("/api/opinions/", methods=["POST"])
 def add_opinion():
     """Добавление отзыва на фильм"""
-    data = request.get_json()
+    data: dict = request.get_json()
+    # Проверка на обязательные поля в data
+    if not ({"title", "text"} <= set(data.keys())):
+        raise InvalidAPIUsage("В запросе отсутствуют обязательные поля")
+    # Проверка на уникальность
+    if Opinion.query.filter_by(text=data["text"]).first() is not None:
+        raise InvalidAPIUsage("Такое мнение уже есть в базе данных")
     # Создание нового пустого отзыва
     opinion = Opinion()
     # Наполнение его данными из запроса
@@ -62,8 +72,10 @@ def add_opinion():
 @app.route("/api/opinions/<int:id>/", methods=["GET"])
 def get_opinion(id):
     """Вывод отзыва по его ID"""
-    # Получить объект по id или выбросить ошибку
-    opinion = Opinion.query.get_or_404(id)
+    opinion = Opinion.query.get(id)
+    if not opinion:
+        raise InvalidAPIUsage("Мнение с указанным id не найдено", 404)
+
     json_string = json.dumps({"opinion": opinion}, ensure_ascii=False)
     # creating a Response object to set the content type and the encoding
     response = Response(
@@ -78,7 +90,16 @@ def get_opinion(id):
 def update_opinion(id):
     """Изменение параметров отзыва"""
     data: dict = request.get_json()
-    opinion: Opinion | None = Opinion.query.get_or_404(id)
+    # Проверка на уникальность текста
+    if (
+        "text" in data
+        and Opinion.query.filter_by(text=data["text"]).first() is not None
+    ):
+        raise InvalidAPIUsage("Такое мнение уже есть в базе данных")
+
+    opinion: Opinion | None = Opinion.query.get(id)
+    if not opinion:
+        raise InvalidAPIUsage("Мнение с указанным id не найдено", 404)
 
     # изменяем атрибуты отзыва
     opinion.update(data)
@@ -98,7 +119,9 @@ def update_opinion(id):
 @app.route("/api/opinions/<int:id>/", methods=["DELETE"])
 def delete_opinion(id):
     """Удаление отзыва"""
-    opinion = Opinion.query.get_or_404(id)
+    opinion: Opinion | None = Opinion.query.get(id)
+    if not opinion:
+        raise InvalidAPIUsage('Мнение с указанным id не найдено', 404)
     db.session.delete(opinion)
     db.session.commit()
